Add files via upload
Ok, so multiple node deletion works, and the algorithm generally works. Here are some issues: The log sigmoid function is weird. Perhaps investigate how it is working. It is not egregious but worth considering. Also, there is a division by zero error in the last loop, in tqi count. The reason has to be blocks not forming the blocks in the last loop, making the number 0. I'm not sure why this is happening, but the rest of it works, so it has to be something small. Only other thing to consider is DBSCAN of the "trash data" we are storing.